requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
cachetools==5.5.2
cattrs==24.1.3
certifi==2025.4.26
charset-normalizer==3.4.2
discord.py==2.3.2
flatbuffers==25.2.10
frozenlist==1.6.0
google-auth==2.40.2
google-genai==1.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jh2==5.0.9
multidict==6.4.4
niquests==3.14.1
numpy==2.2.6
openmeteo_requests==1.5.0
openmeteo_sdk==1.20.0
pandas==2.2.3
platformdirs==4.3.8
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
qh3==1.5.1
requests==2.31.0
requests-cache==1.2.1
retry-requests==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.4.0
urllib3-future==2.12.922
wassima==1.2.2
websockets==15.0.1
yarl==1.20.0
